refactor ints to shorts

+++ Assets/Scripts/MovementHorizontal.cs
@@ -3,11 +3,11 @@ using UnityEngine.Events;
 
 public class MovementHorizontal : MonoBehaviour
 {
-    public int speed;
-    public UnityEvent<int> movementEvent;
+    public short speed;
+    public UnityEvent<short> movementEvent;
 
     public void Trigger(float input)
     {
-        movementEvent.Invoke((int) input * speed);
+        movementEvent.Invoke((short)(input * speed));
     }
 }
